+++ src/test/java/com/caoccao/javet/values/reference/TestV8ValueSymbol.java
@@ -21,21 +21,55 @@ import com.caoccao.javet.BaseTestJavetRuntime;
 import com.caoccao.javet.exceptions.JavetException;
 import org.junit.jupiter.api.Test;
 
-import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.junit.jupiter.api.Assertions.assertNotNull;
+import static org.junit.jupiter.api.Assertions.*;
 
 public class TestV8ValueSymbol extends BaseTestJavetRuntime {
     @Test
-    public void testSymbol() throws JavetException {
-        try (V8ValueSymbol v8ValueSymbol = v8Runtime.getExecutor("Symbol('test')").execute()) {
-            assertNotNull(v8ValueSymbol);
-            assertEquals("test", v8ValueSymbol.getDescription());
-            assertEquals("Symbol(test)", v8ValueSymbol.toString());
+    public void testGlobalSymbol() throws JavetException {
+        try (V8ValueSymbol v8ValueSymbol1 = v8Runtime.createV8ValueSymbol("test", true);
+             V8ValueSymbol v8ValueSymbol2 = v8Runtime.createV8ValueSymbol("test", true);
+             V8ValueSymbol v8ValueSymbol3 = v8Runtime.createV8ValueSymbol("test")) {
+            assertNotNull(v8ValueSymbol1);
+            assertNotNull(v8ValueSymbol2);
+            assertNotNull(v8ValueSymbol3);
+            assertTrue(v8ValueSymbol1.sameValue(v8ValueSymbol2));
+            assertFalse(v8ValueSymbol1.sameValue(v8ValueSymbol3));
+            assertEquals("test", v8ValueSymbol1.getDescription());
+            assertEquals("test", v8ValueSymbol2.getDescription());
+            assertEquals("Symbol(test)", v8ValueSymbol1.toString());
+            assertEquals("Symbol(test)", v8ValueSymbol2.toString());
         }
+    }
+
+    @Test
+    public void testLocalSymbol() throws JavetException {
+        try (V8ValueSymbol v8ValueSymbol1 = v8Runtime.createV8ValueSymbol("test");
+             V8ValueSymbol v8ValueSymbol2 = v8Runtime.createV8ValueSymbol("test")) {
+            assertNotNull(v8ValueSymbol1);
+            assertNotNull(v8ValueSymbol2);
+            assertFalse(v8ValueSymbol1.sameValue(v8ValueSymbol2));
+            assertEquals("test", v8ValueSymbol1.getDescription());
+            assertEquals("test", v8ValueSymbol2.getDescription());
+            assertEquals("Symbol(test)", v8ValueSymbol1.toString());
+            assertEquals("Symbol(test)", v8ValueSymbol2.toString());
+        }
+    }
+
+    @Test
+    public void testNativeNumber() throws JavetException {
         try (V8ValueSymbol v8ValueSymbol = v8Runtime.getExecutor("Symbol(123)").execute()) {
             assertNotNull(v8ValueSymbol);
             assertEquals("123", v8ValueSymbol.getDescription());
             assertEquals("Symbol(123)", v8ValueSymbol.toString());
+        }
+    }
+
+    @Test
+    public void testNativeString() throws JavetException {
+        try (V8ValueSymbol v8ValueSymbol = v8Runtime.getExecutor("Symbol('test')").execute()) {
+            assertNotNull(v8ValueSymbol);
+            assertEquals("test", v8ValueSymbol.getDescription());
+            assertEquals("Symbol(test)", v8ValueSymbol.toString());
         }
     }
 }